# frontend/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Radar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo_user",
              "title": "User Id"
            }
          },
          {
            "name": "unread_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unread Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "title": "Response List Notifications Api V1 Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create Notification",
        "operationId": "create_notification_api_v1_notifications_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create Test Notification",
        "operationId": "create_test_notification_api_v1_notifications_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark All Notifications Read",
        "operationId": "mark_all_notifications_read_api_v1_notifications_read_all_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo_user",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Mark All Notifications Read Api V1 Notifications Read All Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notification",
        "operationId": "get_notification_api_v1_notifications__notification_id__get",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "Update Notification",
        "operationId": "update_notification_api_v1_notifications__notification_id__patch",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Delete Notification",
        "operationId": "delete_notification_api_v1_notifications__notification_id__delete",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/signals/{signal_id}": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Get Signal",
        "operationId": "get_signal_api_v1_signals__signal_id__get",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadarSignal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/{signal_id}/confirm": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Confirm Signal",
        "operationId": "confirm_signal_api_v1_signals__signal_id__confirm_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ManualConfirmRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/{signal_id}/execution-preview": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Preview Virtual Execution",
        "operationId": "preview_virtual_execution_api_v1_signals__signal_id__execution_preview_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ManualConfirmRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualExecutionReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/{signal_id}/reject": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Reject Signal",
        "operationId": "reject_signal_api_v1_signals__signal_id__reject_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ManualRejectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualConfirmRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "virtual",
              "real"
            ],
            "title": "Mode",
            "default": "virtual"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "account_balance": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Account Balance",
            "default": 100.0
          },
          "risk_percent": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Risk Percent",
            "default": 10.0
          },
          "leverage": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Leverage",
            "default": 1
          },
          "size_usd": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Usd"
          },
          "fee_rate": {
            "type": "number",
            "maximum": 0.01,
            "minimum": 0.0,
            "title": "Fee Rate",
            "default": 0.0
          },
          "slippage_bps": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "default": 0.0
          },
          "simulation_mode": {
            "type": "string",
            "enum": [
              "auto",
              "passive",
              "impact_aware"
            ],
            "title": "Simulation Mode",
            "default": "auto"
          },
          "max_virtual_slippage_bps": {
            "type": "number",
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Max Virtual Slippage Bps",
            "default": 150.0
          },
          "allow_partial_fill": {
            "type": "boolean",
            "title": "Allow Partial Fill",
            "default": true
          },
          "min_fill_ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Fill Ratio",
            "default": 0.25
          },
          "market_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VirtualMarketSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_open_positions": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Open Positions",
            "default": 3
          }
        },
        "type": "object",
        "title": "ManualConfirmRequest"
      },
...
      "ManualRejectRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "ManualRejectRequest"
      },
...
      "NotificationCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "type": {
            "type": "string",
            "minLength": 1,
            "title": "Type"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title"
        ],
        "title": "NotificationCreateRequest"
      },
...
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "deliveries": {
            "items": {
              "$ref": "#/components/schemas/NotificationDeliveryResponse"
            },
            "type": "array",
            "title": "Deliveries"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "type",
          "title",
          "body",
          "payload",
          "is_read",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
      "NotificationTestRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "title": "NotificationTestRequest"
      },
      "NotificationUpdateRequest": {
        "properties": {
          "is_read": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Read"
          }
        },
        "type": "object",
        "title": "NotificationUpdateRequest"
      },
...
      "RadarSignal": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "direction": {
            "type": "string",
            "enum": [
              "long",
              "short"
            ],
            "title": "Direction"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "risk_reward": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Reward"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Urgency",
            "default": "medium"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "active",
              "watchlist",
              "confirmed",
              "rejected",
              "expired",
              "invalidated",
              "closed",
              "entry_touched"
            ],
            "title": "Status",
            "default": "active"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "default": 0
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "stream"
          },
          "entry_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Min"
          },
          "entry_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Max"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "take_profit_1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit 1"
          },
          "take_profit_2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit 2"
          },
          "explanation": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanation"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks"
          },
          "score_breakdown": {
            "$ref": "#/components/schemas/SignalScoreBreakdown"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "rejected_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected At"
          },
          "decision_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "virtual",
                  "real"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Mode"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "confirmed_trade_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Trade Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "exchange",
          "strategy",
          "direction",
          "confidence",
          "created_at",
          "updated_at"
        ],
        "title": "RadarSignal"
      },
...
    }
  }
}